+++ bcbio/pipeline/config_utils.py
@@ -238,7 +238,7 @@ def get_jar(base_name, dname):
 # ## Retrieval and update to configuration from arguments
 
 def is_std_config_arg(x):
-    return isinstance(x, dict) and "algorithm" in x and "resources" in x and not "files" in x
+    return isinstance(x, dict) and "algorithm" in x and "resources" in x and "files" not in x
 
 def is_nested_config_arg(x):
     return isinstance(x, dict) and "config" in x and is_std_config_arg(x["config"])