+++ bcbio/provenance/programs.py
@@ -158,22 +158,15 @@ def _get_brew_versions():
     tooldir = install.get_defaults().get("tooldir")
     brew_cmd = os.path.join(tooldir, "bin", "brew") if tooldir else "brew"
     try:
-        vout = subprocess.check_output([brew_cmd, "which"])
-        uses_which = True
-    except subprocess.CalledProcessError:
         vout = subprocess.check_output([brew_cmd, "list", "--versions"])
-        uses_which = False
     except OSError:  # brew not installed/used
         vout = ""
     out = {}
     for vstr in vout.split("\n"):
         if vstr.strip():
-            if uses_which:
-                name, v = vstr.rstrip().split(": ")
-            else:
-                parts = vstr.rstrip().split()
-                name = parts[0]
-                v = parts[-1]
+            parts = vstr.rstrip().split()
+            name = parts[0]
+            v = parts[-1]
             out[name] = v
     return out